getTable moved to table.js
