+++ public/js/table.js
@@ -87,6 +87,44 @@ function getBlankCornerTd() {
     return td;
 }
 
+// TODO: refactor
+function getTable(text) {
+    const table = document.createElement('table');
+    const { charArr, invCharArr } = getCharacterArrays(text); // charArr & invCharArr construction
+
+    // console.log(charArr);
+    // console.log(invCharArr);
+
+    let h = 0;
+    let w = 0;
+    for (let tH = 0; tH < GLOBALS.tableSize; tH += 1) {
+        const tr = document.createElement('tr');
+        if (tH > 0 && tH < 9) w = 0;
+        for (let tW = 0; tW < GLOBALS.tableSize; tW += 1) {
+            let td;
+            if (tW > 0 && tW < 9 && tH > 0 && tH < 9) {
+                // square chars
+                td = getSquareCharTd(charArr[h][w]);
+                w += 1;
+            } else if (tH > 0 && tH < 9) {
+                // vertical dots (left & right)
+                td = getVerticalDotTd(charArr, h, tW);
+            } else if (tW > 0 && tW < 9) {
+                // horizontal dots (top & bottom)
+                td = getHorizontalDotTd(invCharArr, tW, tH);
+            } else {
+                // blank corners
+                td = getBlankCornerTd();
+            }
+            tr.appendChild(td);
+        }
+        table.appendChild(tr);
+        if (tH > 0 && tH < 9) h += 1;
+    }
+
+    return table;
+}
+
 export {
     addToCopy,
     getCharacterArrays,
@@ -94,4 +132,5 @@ export {
     getVerticalDotTd,
     getHorizontalDotTd,
     getBlankCornerTd,
+    getTable,
 };
